@@ -16,7 +16,7 @@
 
 #include "supertux/gameconfig.hpp"
 
-#include "config.h"
+#include <ctime>
 
 #include "editor/overlay_widget.hpp"
 #include "supertux/colorscheme.hpp"
@@ -140,7 +140,7 @@ Config::load()
 
   auto config_mapping = root.get_mapping();
   config_mapping.get("profile", profile);
-  boost::optional<ReaderCollection> config_profiles_mapping;
+  std::optional<ReaderCollection> config_profiles_mapping;
   if (config_mapping.get("profiles", config_profiles_mapping))
   {
     for (auto const& profile_node : config_profiles_mapping->get_objects())
@@ -172,7 +172,7 @@ Config::load()
   config_mapping.get("custom_mouse_cursor", custom_mouse_cursor);
   config_mapping.get("do_release_check", do_release_check);
 
-  boost::optional<ReaderMapping> config_integrations_mapping;
+  std::optional<ReaderMapping> config_integrations_mapping;
   if (config_mapping.get("integrations", config_integrations_mapping))
   {
     config_integrations_mapping->get("hide_editor_levelnames", hide_editor_levelnames);
@@ -181,7 +181,7 @@ Config::load()
 #endif
   }
 
-  boost::optional<ReaderCollection> config_notifications_mapping;
+  std::optional<ReaderCollection> config_notifications_mapping;
   if (config_mapping.get("notifications", config_notifications_mapping))
   {
     for (auto const& notification_node : config_notifications_mapping->get_objects())
@@ -210,7 +210,7 @@ Config::load()
   std::vector<float> menubackcolor_, menufrontcolor_, menuhelpbackcolor_, menuhelpfrontcolor_,
     labeltextcolor_, activetextcolor_, hlcolor_, editorcolor_, editorhovercolor_, editorgrabcolor_;
 
-  boost::optional<ReaderMapping> interface_colors_mapping;
+  std::optional<ReaderMapping> interface_colors_mapping;
   if (config_mapping.get("interface_colors", interface_colors_mapping))
   {
     interface_colors_mapping->get("menubackcolor", menubackcolor_, ColorScheme::Menu::back_color.toVector());
@@ -242,7 +242,7 @@ Config::load()
 
   editor_autotile_help = !developer_mode;
 
-  boost::optional<ReaderMapping> editor_mapping;
+  std::optional<ReaderMapping> editor_mapping;
   if (config_mapping.get("editor", editor_mapping))
   {
     editor_mapping->get("autosave_frequency", editor_autosave_frequency);
@@ -274,7 +274,7 @@ Config::load()
   config_mapping.get("multiplayer_multibind", multiplayer_multibind);
   config_mapping.get("multiplayer_buzz_controllers", multiplayer_buzz_controllers);
 
-  boost::optional<ReaderMapping> config_video_mapping;
+  std::optional<ReaderMapping> config_video_mapping;
   if (config_mapping.get("video", config_video_mapping))
   {
     config_video_mapping->get("fullscreen", use_fullscreen);
@@ -312,7 +312,7 @@ Config::load()
 #endif
   }
 
-  boost::optional<ReaderMapping> config_audio_mapping;
+  std::optional<ReaderMapping> config_audio_mapping;
   if (config_mapping.get("audio", config_audio_mapping))
   {
     config_audio_mapping->get("sound_enabled", sound_enabled);
@@ -321,16 +321,16 @@ Config::load()
     config_audio_mapping->get("music_volume", music_volume);
   }
 
-  boost::optional<ReaderMapping> config_control_mapping;
+  std::optional<ReaderMapping> config_control_mapping;
   if (config_mapping.get("control", config_control_mapping))
   {
-    boost::optional<ReaderMapping> keymap_mapping;
+    std::optional<ReaderMapping> keymap_mapping;
     if (config_control_mapping->get("keymap", keymap_mapping))
     {
       keyboard_config.read(*keymap_mapping);
     }
 
-    boost::optional<ReaderMapping> joystick_mapping;
+    std::optional<ReaderMapping> joystick_mapping;
     if (config_control_mapping->get("joystick", joystick_mapping))
     {
       joystick_config.read(*joystick_mapping);
@@ -340,7 +340,7 @@ Config::load()
     config_control_mapping->get("mobile_controls_scale", m_mobile_controls_scale, 1);
   }
 
-  boost::optional<ReaderCollection> config_addons_mapping;
+  std::optional<ReaderCollection> config_addons_mapping;
   if (config_mapping.get("addons", config_addons_mapping))
   {
     for (auto const& addon_node : config_addons_mapping->get_objects())
@@ -518,6 +518,17 @@ Config::save()
   writer.end_list("editor");
 
   writer.end_list("supertux-config");
+}
+
+
+bool
+Config::is_christmas() const
+{
+  std::time_t time = std::time(nullptr);
+  std::tm* now = std::localtime(&time);
+
+  /* Activate Christmas mode from Dec 6th until Dec 31st. */
+  return now->tm_mday >= 6 && now->tm_mon == 11;
 }
 
 /* EOF */
