@@ -17,7 +17,7 @@
 #ifndef HEADER_SUPERTUX_SUPERTUX_GAMECONFIG_HPP
 #define HEADER_SUPERTUX_SUPERTUX_GAMECONFIG_HPP
 
-#include "config.h"
+#include <optional>
 
 #include "control/joystick_config.hpp"
 #include "control/keyboard_config.hpp"
@@ -25,10 +25,6 @@
 #include "math/vector.hpp"
 #include "video/drawing_context.hpp"
 #include "video/video_system.hpp"
-
-#include <boost/date_time/gregorian/gregorian.hpp>
-#include <boost/date_time/posix_time/posix_time_types.hpp>
-#include <boost/optional.hpp>
 
 class Config final
 {
@@ -88,7 +84,7 @@ public:
   std::string record_demo;
 
   /** this variable is set if tux should spawn somewhere which isn't the "main" spawn point*/
-  boost::optional<Vector> tux_spawn_pos;
+  std::optional<Vector> tux_spawn_pos;
 
   /** force SuperTux language to this locale, e.g. "de". A file
       "data/locale/xx.po" must exist for this to work. An empty string
@@ -157,20 +153,7 @@ public:
 
   std::string repository_url;
 
-  bool is_christmas() const {
-    try
-    {
-      using namespace boost::gregorian;
-      using namespace boost::posix_time;
-      date today = second_clock::local_time().date();
-      date saint_nicholas_day(today.year(), Dec, 6);
-      return today >= saint_nicholas_day;
-    }
-    catch(...)
-    {
-      return false;
-    }
-  }
+  bool is_christmas() const;
 };
 
 #endif
