@@ -26,23 +26,34 @@
 #include "supertux/sector.hpp"
 
 namespace {
+
 const float SNAIL_KICK_SPEED = 500;
 const int MAX_SNAIL_SQUISHES = 10;
 const float SNAIL_KICK_SPEED_Y = -500; /**< y-velocity gained when kicked */
-}
+
+const float SNAIL_GUARD_DELAY = 5.f; /**< Time in-between corrupted snail guard states (seconds) */
+const float SNAIL_GUARD_TIME = 3.f; /**< Duration of corrupted snail guard states (seconds) */
+
+} // namespace
 
 Snail::Snail(const ReaderMapping& reader) :
   WalkingBadguy(reader, "images/creatures/snail/snail.sprite", "left", "right"),
   state(STATE_NORMAL),
   kicked_delay_timer(),
   flat_timer(),
+  m_guard_timer(),
+  m_guard_end_timer(),
   squishcount(0)
 {
+  parse_type(reader);
+  on_type_change(-1);
+
   walk_speed = 80;
   max_drop_height = 600;
   SoundManager::current()->preload("sounds/iceblock_bump.wav");
   SoundManager::current()->preload("sounds/stomp.wav");
   SoundManager::current()->preload("sounds/kick.wav");
+  SoundManager::current()->preload("sounds/dartfire.wav"); // TODO: Specific sounds for snail guard state
 }
 
 void
@@ -52,13 +63,42 @@ Snail::initialize()
   be_normal();
 }
 
+GameObjectTypes
+Snail::get_types() const
+{
+  return {
+    { "normal", _("Normal") },
+    { "corrupted", _("Corrupted") }
+  };
+}
+
+void
+Snail::on_type_change(int old_type)
+{
+  if (!has_found_sprite()) // Change sprite only if a custom sprite has not just been loaded.
+    change_sprite("images/creatures/snail/" + std::string(m_type == Type::CORRUPTED ? "corrupted/corrupted_" : "") + "snail.sprite");
+}
+
 void
 Snail::be_normal()
 {
+  if (m_type == Type::CORRUPTED)
+    m_guard_timer.start(SNAIL_GUARD_DELAY);
+
   if (state == STATE_NORMAL) return;
 
   state = STATE_NORMAL;
   WalkingBadguy::initialize();
+}
+
+void
+Snail::be_guard()
+{
+  if (m_type != Type::CORRUPTED) return;
+
+  state = STATE_GUARD_SHAKE;
+  set_action("shake", m_dir, /* loops = */ 1);
+  m_physic.set_velocity_x(0);
 }
 
 void
@@ -122,17 +162,40 @@ Snail::active_update(float dt_sec)
     return;
   }
 
+  if (m_guard_timer.check() && state == STATE_NORMAL)
+    be_guard();
+
   switch (state) {
 
     case STATE_NORMAL:
       WalkingBadguy::active_update(dt_sec);
       return;
 
+    case STATE_GUARD_SHAKE:
+      if (m_sprite->animation_done())
+      {
+        state = STATE_GUARD;
+        set_action("guard", m_dir);
+        SoundManager::current()->play("sounds/dartfire.wav", get_pos()); // TODO: Specific sounds for snail guard state
+        m_guard_end_timer.start(SNAIL_GUARD_TIME);
+      }
+      break;
+
+    case STATE_GUARD:
+      if (m_guard_end_timer.check())
+      {
+        state = STATE_GUARD_RETRACT;
+        set_action("retract", m_dir, /* loops = */ 1);
+        SoundManager::current()->play("sounds/dartfire.wav", get_pos()); // TODO: Specific sounds for snail guard state
+      }
+      break;
+
     case STATE_FLAT:
       if (flat_timer.check())
         wake_up();
       break;
 
+    case STATE_GUARD_RETRACT:
     case STATE_WAKING:
       if (m_sprite->animation_done())
         be_normal();
@@ -167,6 +230,13 @@ Snail::is_freezable() const
   return true;
 }
 
+bool
+Snail::is_snipable() const
+{
+  return state != STATE_KICKED &&
+         state != STATE_GUARD;
+}
+
 void
 Snail::collision_solid(const CollisionHit& hit)
 {
@@ -192,7 +262,10 @@ Snail::collision_solid(const CollisionHit& hit)
           m_physic.set_velocity_x(-m_physic.get_velocity_x());
         }
       }
-      BOOST_FALLTHROUGH;
+      [[fallthrough]];
+    case STATE_GUARD_SHAKE:
+    case STATE_GUARD:
+    case STATE_GUARD_RETRACT:
     case STATE_FLAT:
     case STATE_KICKED_DELAY:
     case STATE_WAKING:
@@ -218,6 +291,10 @@ Snail::collision_badguy(BadGuy& badguy, const CollisionHit& hit)
   switch (state) {
     case STATE_NORMAL:
       return WalkingBadguy::collision_badguy(badguy, hit);
+    case STATE_GUARD_SHAKE:
+    case STATE_GUARD:
+    case STATE_GUARD_RETRACT:
+      return BadGuy::collision_badguy(badguy, hit);
     case STATE_FLAT:
     case STATE_KICKED_DELAY:
     case STATE_WAKING:
@@ -238,6 +315,9 @@ Snail::collision_player(Player& player, const CollisionHit& hit)
   if (m_frozen)
     return BadGuy::collision_player(player, hit);
 
+  if (state == STATE_GUARD)
+    return WalkingBadguy::collision_player(player, hit); // Hurt player on collision
+
   // handle kicks from left or right side
   if ((state == STATE_WAKING || state == STATE_FLAT) && (hit.left || hit.right)) {
     if (hit.left) {
@@ -256,7 +336,7 @@ Snail::collision_player(Player& player, const CollisionHit& hit)
 bool
 Snail::collision_squished(GameObject& object)
 {
-  if (m_frozen)
+  if (m_frozen || state == STATE_GUARD)
     return WalkingBadguy::collision_squished(object);
 
   Player* player = dynamic_cast<Player*>(&object);
@@ -268,7 +348,7 @@ Snail::collision_squished(GameObject& object)
 
   switch (state) {
     case STATE_NORMAL:
-      BOOST_FALLTHROUGH;
+      [[fallthrough]];
     case STATE_KICKED:
       squishcount++;
       if (squishcount >= MAX_SNAIL_SQUISHES) {
@@ -292,10 +372,9 @@ Snail::collision_squished(GameObject& object)
       }
       be_kicked(true);
       break;
-    case STATE_GRABBED:
-    case STATE_KICKED_DELAY:
-      break;
 
+    default:
+      break;
   }
 
   if (player) player->bounce(*this);
